Added orders and invoices apps

+++ expressmanage/products/models.py
@@ -24,7 +24,7 @@ class ContainerType(TimeStampedModel):
 @with_author
 class RateSlab(TimeStampedModel):
     container_type      = models.ForeignKey(ContainerType, on_delete=models.CASCADE)
-    rate                = models.DecimalField(max_digits=5, decimal_places=2)
+    rate                = models.IntegerField()
     number_of_days      = models.IntegerField()
     slab_number         = models.IntegerField()
